+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 This release adds support to OneView Rest API version 500 for the hardware variants C7000 and Synergy to the already existing features:
    - Drive Enclosure
    - Enclosure
+   - Enclosure Group
    - Ethernet Network
    - FC Network
    - FCoE Network